requirements.txt
# Core runtime dependencies
fastapi>=0.100.0
httpx>=0.24.0
pyyaml>=6.0
uvicorn>=0.22.0
click>=8.0
